requirements.txt
PyTelegramBotAPI~=4.25.0

python-dotenv~=1.0.1